+++ usemaps_lite/tools/organization.py
@@ -1,13 +1,18 @@
 from typing import Dict, Any
 
 from PyQt5 import QtWidgets
-from PyQt5.QtGui import QStandardItem
+from PyQt5.QtGui import QIcon
+from PyQt5.Qt import QStandardItem
 from qgis.PyQt.QtCore import Qt
+from PyQt5.QtWidgets import QMessageBox
 
 from usemaps_lite.tools.base_logic_class import BaseLogicClass
 from usemaps_lite.ui.invite_user import InviteUserDialog
 from usemaps_lite.tools.event_handler import Event
 from usemaps_lite.tools.user_mapper import USER_MAPPER
+from usemaps_lite.tools.metadata import ORGANIZATION_METADATA
+from usemaps_lite.tools.translations import TRANSLATOR
+
 
 class Organization(BaseLogicClass):
     """
@@ -25,9 +30,12 @@ class Organization(BaseLogicClass):
         self.event_handler.register_event_handler(Event.VERIFIED_USER, self.handle_verified_user_event)
         self.event_handler.register_event_handler(Event.DELETED_USER, self.handle_deleted_user_event)
         self.event_handler.register_event_handler(Event.NEW_COMMENT, self.handle_new_comment_event)
+        self.event_handler.register_event_handler(Event.ONLINE_USER, self.handle_online_user_event)
+        self.event_handler.register_event_handler(Event.OFFLINE_USER, self.handle_offline_user_event)
 
         self.dockwidget.users_tableview.setColumnWidth(0, 200)
         self.dockwidget.users_tableview.setColumnWidth(1, 100)
+        self.dockwidget.users_tableview.setColumnWidth(2, 100)
         self.dockwidget.users_tableview.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
         self.dockwidget.users_tableview.selectionModel().selectionChanged.connect(self.on_users_tableview_selection_changed)
 
@@ -72,7 +80,7 @@ class Organization(BaseLogicClass):
         """
 
         if (error_msg := response.get("error")) is not None:
-            self.show_error_message(f"Błąd dodawania komentarza: {error_msg}")
+            self.show_error_message(f"{TRANSLATOR.translate_error('comment')}: {error_msg}")
 
         else:
             self.dockwidget.comment_lineedit.clear()
@@ -97,10 +105,15 @@ class Organization(BaseLogicClass):
         """
 
         if (error_msg := response.get("error")) is not None:
-            self.show_error_message(f"Błąd wysyłania zaproszenia: {error_msg}")
+            
+            if error_msg.get('server_message') == 'user already exists':
+                self.show_error_message(TRANSLATOR.translate_error("user exists"))
+
+            else:
+                self.show_error_message(f"{TRANSLATOR.translate_error('invite')}: {error_msg.get('server_message')}")
 
         else:
-            self.show_success_message("Wysłano zaproszenie.")
+            self.show_success_message(TRANSLATOR.translate_info("invite send"))
             self.invite_user_dialog.hide()
 
 
@@ -109,22 +122,32 @@ class Organization(BaseLogicClass):
         Wykonuje request usunięcia wybranego członka organizacji
         """
 
-        selected_index = self.dockwidget.users_tableview.selectedIndexes()
-        user_email = selected_index[0].data()
-
-        self.api.delete(
-            "org/users",
-            {"email": user_email},
-            callback=self.handle_delete_user_response
+        reply = QMessageBox.question(
+            self.dockwidget,
+            TRANSLATOR.translate_ui("remove user label"),
+            TRANSLATOR.translate_ui("remove user question"),
+            QMessageBox.Yes | QMessageBox.No,
+            QMessageBox.No
         )
-    
+
+        if reply == QMessageBox.Yes:
+
+            selected_index = self.dockwidget.users_tableview.selectedIndexes()
+            user_email = selected_index[0].data()
+
+            self.api.delete(
+                "org/users",
+                {"email": user_email},
+                callback=self.handle_delete_user_response
+            )
+        
     def handle_delete_user_response(self, response: Dict[str, Any]) -> None:
         """
         Obsługuje odpowiedź po próbie usunięcia członka organizacji
         """
 
         if (error_msg := response.get("error")) is not None:
-            self.show_error_message(f"Błąd usuwania użytkownika: {error_msg}")
+            self.show_error_message(f"{TRANSLATOR.translate_error('remove user')}: {error_msg}")
 
     def on_users_tableview_selection_changed(self) -> None:
         """
@@ -139,28 +162,28 @@ class Organization(BaseLogicClass):
         """
         Obsługuje przychodzące zdarzenie zaproszenia nowego członka do organizacji.
         """
-        
+
         data = event_data.get("data")
-        user_uuid_who_invited = event_data.get("user")
-        
+
         email = data.get("email")
         user_uuid = data.get("uuid")
         verified = "Tak" if data.get("verified") else "Nie"
-        
+
         email_item = QStandardItem(email)
         email_item.setData(user_uuid, Qt.UserRole)
-        
+
+        icon = QIcon(":images/themes/default/repositoryDisabled.svg")
+        icon_item = QStandardItem()
+        icon_item.setIcon(icon)
+
         row = [
             email_item,
-            QStandardItem(verified)
+            QStandardItem(verified),
+            icon_item
         ]
-        
+
         self.dockwidget.users_tableview_model.appendRow(row)
-
-        user_email_who_invited = USER_MAPPER.get_user_email(user_uuid_who_invited)
-        
-        self.show_info_message(f"Użytkownik {user_email_who_invited} zaprosił użytkownika {email}")
-
+        self.toggle_invite_user_button()
 
     def handle_verified_user_event(self, event_data: Dict[str, Any]) -> None:
         """
@@ -174,9 +197,11 @@ class Organization(BaseLogicClass):
             item = self.dockwidget.users_tableview_model.item(user_row, 0)
             if item and item.text() == email:
                 verified_item = QStandardItem("Tak")
+                icon = QIcon(":images/themes/default/repositoryConnected.svg")
+                icon_item = QStandardItem()
+                icon_item.setIcon(icon)
                 self.dockwidget.users_tableview_model.setItem(user_row, 1, verified_item)
-        
-        self.show_info_message(f"Użytkownik {email} zweryfikował konto")
+                self.dockwidget.users_tableview_model.setItem(user_row, 2, icon_item)
 
     def handle_deleted_user_event(self, event_data: Dict[str, Any]) -> None:
         """
@@ -185,7 +210,6 @@ class Organization(BaseLogicClass):
 
         data = event_data.get("data")
         user_uuid = data.get("uuid")
-        user_email = data.get("email")
 
         row_to_remove = -1
 
@@ -197,18 +221,66 @@ class Organization(BaseLogicClass):
 
         if row_to_remove != -1:
             self.dockwidget.users_tableview_model.removeRow(row_to_remove)
-        
-        user_uuid_who_removed = event_data.get("user")
-        user_email_who_removed = USER_MAPPER.get_user_email(user_uuid_who_removed)
-        
-        self.show_info_message(f"Użytkownik {user_email_who_removed} usunął użytkownika {user_email}")
+
+        self.toggle_invite_user_button()
 
     def handle_new_comment_event(self, event_data: Dict[str, Any]) -> None:
         """
         Obsługuje przychodzące zdarzenie nowego komentarza.
         """
-        
+
         user_uuid = event_data.get("user")
         user_email = USER_MAPPER.get_user_email(user_uuid)
+        self.show_info_message(f"{user_email} {TRANSLATOR.translate_info('added new comment')}")
+
+    def toggle_invite_user_button(self):
+        """
+        Włącza/wyłącza przycisk zapraszania współpracowników, w zaleznosci od osiągnięcia
+        limitu współpracowników w organizacji
+        """
         
-        self.show_info_message(f"Użytkownik {user_email} dodał nowy komentarz")
+        num_of_users_limit = ORGANIZATION_METADATA.get_num_of_users_limit()
+        num_of_users_added = self.dockwidget.users_tableview_model.rowCount()
+        
+        self.dockwidget.invite_user_button.setEnabled(num_of_users_added < num_of_users_limit)
+
+    def handle_online_user_event(self, event_data: Dict[str, Any]) -> None:
+        """
+        Obsługuje przychodzące zdarzenie zalogowania współpracownika
+        """
+
+        data = event_data.get("data")
+        online_user_uuid = data.get("uuid")
+        online_user_email = USER_MAPPER.get_user_email(online_user_uuid)
+
+        for user_row in range(self.dockwidget.users_tableview_model.rowCount()):
+            item = self.dockwidget.users_tableview_model.item(user_row, 0)
+            if item and item.text() == online_user_email:
+                icon = QIcon(":images/themes/default/repositoryConnected.svg")
+                icon_item = QStandardItem()
+                icon_item.setIcon(icon)
+                self.dockwidget.users_tableview_model.setItem(user_row, 2, icon_item)
+
+                
+        self.show_info_message(f"{online_user_email} {TRANSLATOR.translate_info('is online')}")
+
+
+    def handle_offline_user_event(self, event_data: Dict[str, Any]) -> None:
+        """
+        Obsługuje przychodzące zdarzenie zalogowania współpracownika
+        """
+
+        data = event_data.get("data")
+        online_user_uuid = data.get("uuid")
+        online_user_email = USER_MAPPER.get_user_email(online_user_uuid)
+
+        for user_row in range(self.dockwidget.users_tableview_model.rowCount()):
+            item = self.dockwidget.users_tableview_model.item(user_row, 0)
+            if item and item.text() == online_user_email:
+                icon = QIcon(":images/themes/default/repositoryDisabled.svg")
+                icon_item = QStandardItem()
+                icon_item.setIcon(icon)
+                self.dockwidget.users_tableview_model.setItem(user_row, 2, icon_item)
+
+        self.show_info_message(f"{online_user_email} {TRANSLATOR.translate_info('is offline')}")
+
